be/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.22
httpx>=0.24
python-dotenv>=1.0
pydantic>=2.0
